requirements.txt
discord.py==2.6.4
requests>=2.31.0
python-dotenv